@@ -2,10 +2,20 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
 const routes: Routes = [
-  { path: '', pathMatch: 'full', redirectTo: 'home' },
-  {path: 'home',loadChildren: () => import('./home/home.module').then((m) => m.HomeModule),
+  {
+    path: '',
+    pathMatch: 'full',
+    redirectTo: 'home',
   },
-  {path: 'animais',loadChildren: () => import('./animais/animais.module').then((m) => m.AnimaisModule)},
+  {
+    path: 'home',
+    loadChildren: () => import('./home/home.module').then((m) => m.HomeModule),
+  },
+  {
+    path: 'animais',
+    loadChildren: () =>
+      import('./animais/animais.module').then((m) => m.AnimaisModule),
+  },
 ];
 
 @NgModule({
