@@ -20,16 +20,16 @@ describe('AppComponent', () => {
     expect(app).toBeTruthy();
   });
 
-  it(`should have as title 'gatitoBook'`, () => {
+  it(`should have as title 'gatitobook'`, () => {
     const fixture = TestBed.createComponent(AppComponent);
     const app = fixture.componentInstance;
-    expect(app.title).toEqual('gatitoBook');
+    expect(app.title).toEqual('gatitobook');
   });
 
   it('should render title', () => {
     const fixture = TestBed.createComponent(AppComponent);
     fixture.detectChanges();
     const compiled = fixture.nativeElement;
-    expect(compiled.querySelector('.content span').textContent).toContain('gatitoBook app is running!');
+    expect(compiled.querySelector('.content span').textContent).toContain('gatitobook app is running!');
   });
 });
