@@ -3,13 +3,13 @@ import { Component, Input, OnInit } from '@angular/core';
 @Component({
   selector: 'app-mensagem',
   templateUrl: './mensagem.component.html',
-  styleUrls: ['./mensagem.component.css']
+  styleUrls: ['./mensagem.component.css'],
 })
 export class MensagemComponent implements OnInit {
-  @Input() mensagem: string = '';
-  constructor() { }
+  @Input()
+  mensagem = '';
 
-  ngOnInit(): void {
-  }
+  constructor() {}
 
+  ngOnInit(): void {}
 }
