@@ -2,11 +2,9 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { MensagemComponent } from './mensagem.component';
 
-
-
 @NgModule({
   declarations: [MensagemComponent],
   imports: [CommonModule],
-  exports: [MensagemComponent]
+  exports: [MensagemComponent],
 })
-export class MensagemModule { }
+export class MensagemModule {}
