@@ -1,5 +1,3 @@
-import { AnimaisModule } from './../animais/animais.module';
-import { MensagemModule } from './../componentes/mensagem/mensagem.module';
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
@@ -7,8 +5,8 @@ import { HomeRoutingModule } from './home-routing.module';
 import { HomeComponent } from './home.component';
 import { LoginComponent } from './login/login.component';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { MensagemModule } from '../componentes/mensagem/mensagem.module';
 import { NovoUsuarioComponent } from './novo-usuario/novo-usuario.component';
-
 
 @NgModule({
   declarations: [HomeComponent, LoginComponent, NovoUsuarioComponent],
@@ -17,9 +15,8 @@ import { NovoUsuarioComponent } from './novo-usuario/novo-usuario.component';
     HomeRoutingModule,
     FormsModule,
     MensagemModule,
-    AnimaisModule,
-    ReactiveFormsModule
+    ReactiveFormsModule,
   ],
   exports: [HomeComponent],
 })
-export class HomeModule { }
+export class HomeModule {}
