@@ -1,26 +1,32 @@
-import { AutenticacaoService } from './../../autenticacao/autenticacao.service';
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
+import { AutenticacaoService } from 'src/app/autenticacao/autenticacao.service';
 
 @Component({
   selector: 'app-login',
   templateUrl: './login.component.html',
-  styleUrls: ['./login.component.css']
+  styleUrls: ['./login.component.css'],
 })
 export class LoginComponent implements OnInit {
-  usuario: string = '';
-  senha: string = '';
-  constructor(private authService: AutenticacaoService, private router: Router) { }
+  usuario = '';
+  senha = '';
 
-  ngOnInit(): void {
-  }
+  constructor(
+    private authService: AutenticacaoService,
+    private router: Router
+  ) {}
 
-  login(){
-    this.authService.autenticar(this.usuario, this.senha).subscribe(() => {
-      console.log('Autentica com sucesso');
-      this.router.navigate(['animais']);
-    }, err => {
-      alert('Usuário ou senha inválidos');
-    });
+  ngOnInit(): void {}
+
+  login() {
+    this.authService.autenticar(this.usuario, this.senha).subscribe(
+      () => {
+        this.router.navigate(['animais']);
+      },
+      (error) => {
+        alert('Usuário ou senha inválido');
+        console.log(error);
+      }
+    );
   }
 }
