@@ -6,21 +6,19 @@ import { first, map, switchMap } from 'rxjs/operators';
 @Injectable({
   providedIn: 'root',
 })
-export class UsuarioExisteService {
-  constructor(private novoUsuarioSvc: NovoUsuarioService) {}
+export class UsuarioExistenteService {
+  constructor(private novoUsuarioSvc: NovoUsuarioService, ) {}
 
-  usuarioJaexiste() {
+  usuarioJaExiste(){
     return (control: AbstractControl) => {
-      return (
-        control.valueChanges.pipe(
-          switchMap((nomeUsuario) =>
-            this.novoUsuarioSvc.verificaUsuarioExistente(nomeUsuario)
-          )
-        ),
-        map((usuarioExiste) =>
-          usuarioExiste ? { usuarioExistente: true } : null
-        ),
-        first()
+      return control.valueChanges.pipe(
+        switchMap((nomeUsuario) =>
+          this.novoUsuarioSvc.verificaUsuarioExistente(nomeUsuario)
+      ),
+      map((usuarioExiste) =>
+      usuarioExiste ? { usuarioExistente:true } : null
+      ),
+      first()
       );
     };
   }
